moved reduce step to its own script:   tests/test_workflow.py
